Refactor jb::itch5::testing to its own directory
