+++ jb/itch5/ut_compute_inside.cpp
@@ -1,4 +1,5 @@
-#include <jb/itch5/compute_inside.hpp>
+#include "jb/itch5/compute_inside.hpp"
+#include <jb/itch5/testing/messages.hpp>
 #include <jb/as_hhmmss.hpp>
 
 #include <skye/mock_function.hpp>
@@ -28,27 +29,6 @@ buy_sell_indicator_t const SELL(u'S');
 timestamp create_timestamp() {
   return timestamp{std::chrono::nanoseconds(0)};
 }
-
-/// Create a stock_directory_message for testing
-stock_directory_message create_stock_directory(char const* symbol) {
-  return stock_directory_message{
-      {stock_directory_message::message_type, 0, 0, create_timestamp()},
-      stock_t(symbol),
-      market_category_t(u'Q'),
-      financial_status_indicator_t(u'N'),
-      100,
-      roundlots_only_t('N'),
-      issue_classification_t(u'C'),
-      issue_subtype_t("C"),
-      authenticity_t(u'P'),
-      short_sale_threshold_indicator_t(u' '),
-      ipo_flag_t(u'N'),
-      luld_reference_price_tier_t(u' '),
-      etp_flag_t(u'N'),
-      0,
-      inverse_indicator_t(u'N')};
-}
-
 } // anonymous namespace
 
 /**
@@ -77,6 +57,7 @@ BOOST_AUTO_TEST_CASE(compute_inside_simple) {
 
   // ... send a couple of stock directory messages, do not much care
   // about their contents other than the symbol ...
+  using namespace jb::itch5::testing;
   compute_inside::time_point now = tested.now();
   long msgcnt = 0;
   tested.handle_message(now, ++msgcnt, 0, create_stock_directory("HSART"));
