+++ jb/itch5/ut_message_header.cpp
@@ -1,5 +1,5 @@
 #include <jb/itch5/message_header.hpp>
-#include <jb/itch5/testing_data.hpp>
+#include <jb/itch5/testing/data.hpp>
 
 #include <boost/test/unit_test.hpp>
