+++ jb/itch5/ut_mold_udp_pacer.cpp
@@ -1,5 +1,5 @@
 #include <jb/itch5/mold_udp_pacer.hpp>
-#include <jb/itch5/testing_data.hpp>
+#include <jb/itch5/testing/data.hpp>
 
 #include <skye/mock_function.hpp>
 #include <boost/array.hpp>
