+++ jb/itch5/ut_order_cancel_message.cpp
@@ -1,5 +1,5 @@
 #include <jb/itch5/order_cancel_message.hpp>
-#include <jb/itch5/testing_data.hpp>
+#include <jb/itch5/testing/data.hpp>
 
 #include <boost/test/unit_test.hpp>
